moving FilePicker and Editor components out of old-editor directory, so that there are no longer any dependencies on the old-editor directory
