+++ app/javascript/components/workflows/professor/exams/new/editor/components/FilePicker.tsx
