+++ app/javascript/components/workflows/professor/exams/new/old-editor/components/questions/Prompted.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react';
 import { Form, Row, Col } from 'react-bootstrap';
-import CustomEditor from '@professor/exams/new/old-editor/components/CustomEditor';
+import CustomEditor from '@professor/exams/new/editor/components/CustomEditor';
 import { WrappedFieldProps, Field } from 'redux-form';
 
 const EditPrompt: React.FC<WrappedFieldProps & {
